fix: moved MobileDatabase class to utils
This class seems to be more an util than an
main class. Now is located at mobile package utils
