+++ Sensorable/mobile/src/main/java/com/sensorable/activities/BluetoothOptionsActivity.java
@@ -14,7 +14,7 @@ import com.commons.devicesDetection.BluetoothDevicesProvider;
 import com.commons.SensorableConstants;
 import com.commons.database.BluetoothDeviceDao;
 import com.sensorable.utils.BluetoothDeviceAdapter;
-import com.sensorable.MobileDatabase;
+import com.sensorable.utils.MobileDatabase;
 import com.sensorable.R;
 
 import java.util.ArrayList;
